+++ foxcross/exceptions.py
@@ -3,8 +3,16 @@ class NoModelServingFoundError(Exception):
 
 
 class BadDataFormatError(Exception):
-    pass
+    http_status_code = 400
 
 
 class TestDataPathUndefinedError(Exception):
     pass
+
+
+class PreProcessingError(Exception):
+    http_status_code = 400
+
+
+class PostProcessingError(Exception):
+    http_status_code = 500