+++ tests/test_performance_extra.py
@@ -1,5 +1,0 @@
-def test_pandas_extra():
-    import modin  # noqa: F401
-    import pandas  # noqa: F401
-    import numpy  # noqa: F401
-    import ujson  # noqa: F401